Fix some compile warnings and general clean up (From Robbert Muller)

+++ ircd/s_bsd_ext.h
@@ -37,7 +37,7 @@ extern time_t timeofday;
 #else /* S_BSD_C */
 #define EXTERN
 #endif /* S_BSD_C */
-EXTERN void add_local_domain __P((char *hname, int size));
+EXTERN void add_local_domain __P((char *hname, size_t size));
 EXTERN void report_error __P((char *text, aClient *cptr));
 EXTERN int inetport __P((aClient *cptr, char *ip, char *ipmask, int port));
 EXTERN int add_listener __P((aConfItem *aconf));
